+++ BitSim/BitSim/BitmexSimulator.h
@@ -11,11 +11,11 @@ public:
         intervals(intervals),
         intervals_idx_start(0), intervals_idx_end(0),
         intervals_idx(0),
-        wallet(0.0f), entry_price(0.0f), pos_contracts(0.0f) {}
+        wallet(0.0), entry_price(0.0), pos_contracts(0.0) {}
 
     void reset(void);
 
-    void put_order(float price, float contracts);
+    void put_order(double price, double contracts);
 
 private:
     sptrIntervals intervals;
@@ -24,11 +24,11 @@ private:
     int intervals_idx_end;
     int intervals_idx;
 
-    float wallet;
-    float entry_price;
-    float pos_contracts;
+    double wallet;
+    double entry_price;
+    double pos_contracts;
 
-    void execute_order(float price, float contracts, bool taker);
+    void execute_order(double price, double contracts, bool taker);
 };
 
 using sptrBitmexSimulator = std::shared_ptr<BitmexSimulator>;
